requirements.txt
streamlit
openai
python-docx
scikit-learn